Feature/timer

+++ lib/sys/include/common.h
@@ -28,8 +28,6 @@
 #define uint8_t  unsigned char
 #define bool_t   bool
 
-// #define NULL 0
-
 /******************************************************************************
  * @struct k_return_t
  * @brief general kernel error return code
@@ -55,7 +53,7 @@ inline void reg_write_double_word(const uint64_t addr, const uint64_t data) {
  * @param addr of the register
  * @return data in the register
  ******************************************************************************/
-inline uint8_t reg_read_double_word(const uint64_t addr) {
+inline uint64_t reg_read_double_word(const uint64_t addr) {
   volatile uint64_t *reg_addr = (uint64_t *)addr;
   return (uint64_t)*reg_addr;
 }
